Rollback TrayMenuController to FxApplication code moving
